LLM-DA++ for TPR/requirements.txt
timm==0.6.11
wandb==0.13.5
ftfy==6.1.1

regex
easydict
pyyaml
textaugment
ipdb

torchmetrics
matplotlib
jupyter
ipykernel
